reconfigured unit tests
Unit tests previously relied on using the application's default database. This caused issues when the application switched from an in-memory Hibernate DB to an AWS RDS DB. This commit migrates the unit tests to rely on a MySQL testcontainer, as well as rearranging the unit tests by extracting common functionality.